add midleware and new endpoint for signed urls

+++ src/controllers/bidDocumentController.js
@@ -1,9 +1,15 @@
 const path = require('path');
 const crypto = require('crypto');
+
 const supabaseBucketService = require("../services/supabaseBucketService");
 const bidDocumentRepository = require("../repositories/bidDocumentRepository");
+const bidDocumentService = require("../services/bidDocumentService");
 
 exports.uploadBidDocument = async (req, res) => {
+    if (req.fileValidationError) {
+        return res.status(400).json({ message: req.fileValidationError });
+    }
+
     const file = req.file;
     if (!file) {
         return res.status(400).json({ message: "No file uploaded" });
@@ -37,6 +43,10 @@ exports.deleteBidDocument = async (req, res) => {
     }
 
     const bidDocument = await bidDocumentRepository.getBidDocument(id);
+    if (!bidDocument) {
+        return res.status(404).json({ message: "Document not found" });
+    }
+
     const filePath = bidDocument.file_path;
 
     const removedCount = await bidDocumentRepository.removeBidDocument(id);
@@ -47,4 +57,18 @@ exports.deleteBidDocument = async (req, res) => {
     const deleted = await supabaseBucketService.deleteFile(filePath);
 
     return res.status(200).json({ message: "File deleted successfully" });
+}
+
+exports.getBidDocumentsByProcurementBidId = async (req, res) => {
+    const { id } = req.params;
+    if (!id) {
+        return res.status(400).json({ message: "No ID provided" });
+    }
+
+    const bidDocuments = await bidDocumentService.getBidDocumentsByProcurementBidId(id);
+    if (!bidDocuments) {
+        return res.status(404).json({ message: "No documents found" });
+    }
+
+    return res.status(200).json(bidDocuments);
 }
